Stabilize scripts up to stage 2 GCC
The scripts worked for the most part, but would break if you tried to
simply run them sequentially. They should now work from the get-go.  You
still need to look at the top of the scripts to see if you need to run a
patch or not. I'm not running them automatically because I'm not
confident enough that they'll always work on the first try every time
for everyone.